# docs/api/openapi-agent-centric-draft.yaml
openapi: 3.1.0
info:
  title: Master Agent API (Agent-Centric Draft)
  version: 0.1.0-draft
  description: Draft contract for multi-provider multi-agent orchestration.
servers:
  - url: http://localhost:3000
paths:
  /api/v1/agents:
    get:
      summary: List provider agents
      parameters:
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Agent list page
    post:
      summary: Create or enable provider agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created
  /api/v1/agents/{agentId}:
    get:
      summary: Get agent detail
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent detail
  /api/v1/agents/{agentId}/start:
    post:
      summary: Start agent execution
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Start accepted
  /api/v1/agents/{agentId}/stop:
    post:
      summary: Stop running agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stop accepted
  /api/v1/agents/{agentId}/followup:
    post:
      summary: Send follow-up prompt to agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowupRequest"
      responses:
        "200":
          description: Follow-up accepted
  /api/v1/agents/{agentId}/messages:
    get:
      summary: Paginated agent messages
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Message page
  /api/v1/agents/{agentId}/logs:
    get:
      summary: Paginated agent logs
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Log page
  /api/v1/agents/{agentId}/file-changes:
    get:
      summary: Paginated file change history
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: File changes page
  /api/v1/providers/{provider}/sync:
    post:
      summary: Trigger provider sync job
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Sync started
components:
  schemas:
    CreateAgentRequest:
      type: object
      required: [provider, name]
      properties:
        provider:
          type: string
          enum: [cursor_cloud, cursor_cli, anthropic, codex]
        name:
          type: string
        projectId:
          type: string
        config:
          type: object
    FollowupRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string